Add back Chart.Ticks.formatters (#5088)
